sbi-benchmark/requirements_koopman.txt
# Requirements for Koopman SBI Implementation
# Install with: pip install -r requirements_koopman.txt

# Core dependencies
torch>=1.12.0
numpy>=1.21.0
PyYAML>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.4.0

# SBI benchmark
sbibm>=1.0.0

# Scientific computing
scipy>=1.7.0
scikit-learn>=1.1.0

# Optional utilities
tqdm>=4.64.0
wandb>=0.12.0

# Development/testing
pytest>=7.0.0